ui: remove the classic design — the workbench is the only UI
Founder decision, 2026-08-30: "I don't think we need to keep the old design."

Gone: the v1 `App` and its chat / vault / settings components, the `ui`
design flag (`ui-flag.ts`, `localStorage['counsel-os.ui']`, `#/?ui=v1`,
`<html data-ui>`), the Settings "New design" switch, and the v1 e2e story.
`app.tsx` keeps only the fragment router, TOKEN_MESSAGE, and a `Root` that
mounts `Shell`.

The v2 tokens in `styles.css` were scoped to `html[data-ui="v2"]`; they are
merged into `:root` and are now the page's only palette. v1-only component
CSS is deleted; the classes shared components still draw (`.markdown`,
`.vault-*`, `.facts`, `.providers-table`, `.field`, `.provider-test-*`,
`.notice`, `.badge`, `.page-message`) stay. The `.v2-*` prefix is kept —
renaming it would be a diff with no reader.

`settings/ProvidersForm.tsx` went too: only the v1 Settings page rendered
it. Its test was the only cover for `ProviderTest`'s confirm, cancel and
failure paths, and `ProviderTest` is still shared — so those three cases
move to a new `settings/ProviderTest.test.tsx` rather than being lost.

The one e2e story now opens the printed URL and asserts the workbench. With
no v1 story ahead of it the vault starts with no threads, so the shell opens
its own draft: the step asserts that (and that "New" is disabled) instead of
clicking "New".
